fix: ESLint stops linting on first error with flat config
