+++ lib/eslint/eslint.js
@@ -16,6 +16,7 @@ const path = require("path");
 const findUp = require("find-up");
 const { version } = require("../../package.json");
 const { Linter } = require("../linter");
+const { onFatalError } = require("../shared/error-utils");
 const { getRuleFromConfig } = require("../config/flat-config-helpers");
 const {
     Legacy: {
@@ -919,8 +920,12 @@ class ESLint {
                         }
 
                         return result;
-                    });
+                    }).catch(error => {
+                        onFatalError(error);
 
+                        // eslint-disable-next-line n/no-process-exit -- Intentionally exit to avoid running the process further.
+                        process.exit();
+                    });
             })
         );
